+++ playgroundsound/playgroundsound/WwiseAPI.cpp
@@ -105,13 +105,13 @@ std::vector<DiffractionPath> WwiseAPI::GetDiffraction(const AkGameObjectID& game
 		AkDiffractionPathInfo akDiffractionPath = akDiffractionPaths[path];
 		for (size_t i = 0; i < akDiffractionPath.nodeCount; i++)
 		{
-			diffractionPath.nodes[i].x = akDiffractionPath.nodes[i].X;
-			diffractionPath.nodes[i].y = akDiffractionPath.nodes[i].Y;
-			diffractionPath.nodes[i].z = -akDiffractionPath.nodes[i].Z;
+			diffractionPath.nodes[i].x = static_cast<float>(akDiffractionPath.nodes[i].X);
+			diffractionPath.nodes[i].y = static_cast<float>(akDiffractionPath.nodes[i].Y);
+			diffractionPath.nodes[i].z = static_cast<float>(-akDiffractionPath.nodes[i].Z);
 		}
-		diffractionPath.emitterPos.x = emitterPos.X;
-		diffractionPath.emitterPos.y = emitterPos.Y;
-		diffractionPath.emitterPos.z = emitterPos.Z;
+		diffractionPath.emitterPos.x = static_cast<float>(emitterPos.X);
+		diffractionPath.emitterPos.y = static_cast<float>(emitterPos.Y);
+		diffractionPath.emitterPos.z = static_cast<float>(emitterPos.Z);
 		for (size_t i = 0; i < akDiffractionPath.nodeCount; i++)
 		{
 			diffractionPath.angles[i] = akDiffractionPath.angles[i];
@@ -119,12 +119,12 @@ std::vector<DiffractionPath> WwiseAPI::GetDiffraction(const AkGameObjectID& game
 		for (size_t i = 0; i < akDiffractionPath.nodeCount; i++)
 		{
 
-			diffractionPath.portals[i] = akDiffractionPath.portals[i];
+			diffractionPath.portals[i] = static_cast<int>(akDiffractionPath.portals[i]);
 		}
 		for (size_t i = 0; i < akDiffractionPath.nodeCount; i++)
 		{
 
-			diffractionPath.rooms[i] = akDiffractionPath.rooms[i];
+			diffractionPath.rooms[i] = static_cast<int>(akDiffractionPath.rooms[i]);
 		}
 		diffractionPath.virtualPos.forward.x = -akDiffractionPath.virtualPos.OrientationFront().X;
 		diffractionPath.virtualPos.forward.y = -akDiffractionPath.virtualPos.OrientationFront().Y;
@@ -132,9 +132,9 @@ std::vector<DiffractionPath> WwiseAPI::GetDiffraction(const AkGameObjectID& game
 		diffractionPath.virtualPos.up.x = akDiffractionPath.virtualPos.OrientationTop().X;
 		diffractionPath.virtualPos.up.y = akDiffractionPath.virtualPos.OrientationTop().Y;
 		diffractionPath.virtualPos.up.z = -akDiffractionPath.virtualPos.OrientationTop().Z;
-		diffractionPath.virtualPos.position.x = akDiffractionPath.virtualPos.Position().X;
-		diffractionPath.virtualPos.position.y = akDiffractionPath.virtualPos.Position().Y;
-		diffractionPath.virtualPos.position.z = -akDiffractionPath.virtualPos.Position().Z;
+		diffractionPath.virtualPos.position.x = static_cast<float>(akDiffractionPath.virtualPos.Position().X);
+		diffractionPath.virtualPos.position.y = static_cast<float>(akDiffractionPath.virtualPos.Position().Y);
+		diffractionPath.virtualPos.position.z = static_cast<float>(-akDiffractionPath.virtualPos.Position().Z);
 		diffractionPath.nodeCount = akDiffractionPath.nodeCount;
 		diffractionPath.diffraction = akDiffractionPath.diffraction;
 		diffractionPath.totLength = akDiffractionPath.totLength;
